Move `CanvasTextUtils` out from `EuiTextTruncate` to `services/`
- separating it from EuiTextTruncate utils makes it easier to mock without shenanigans

- since it's being used outside of truncation contexts, might as well make it a generic service
