+++ src/components/text_truncate/utils.ts
@@ -6,71 +6,13 @@
  * Side Public License, v 1.
  */
 
-import type { ExclusiveUnion } from '../common';
+import { CanvasTextParams, CanvasTextUtils } from '../../services/canvas';
 
-type CanvasTextParams = ExclusiveUnion<
-  { container: HTMLElement },
-  { font: CanvasTextDrawingStyles['font'] }
->;
 type TruncationParams = CanvasTextParams & {
   fullText: string;
   ellipsis: string;
   availableWidth: number;
 };
-
-/**
- * Under the hood, a temporary Canvas element is created for manipulating text
- * & determining text width.
- *
- * To accurately measure text, canvas rendering requires either a container to
- * compute/derive font styles from, or a static font string (useful for usage
- * outside the DOM). Particular care should be applied when fallback fonts are
- * used, as more fallback fonts can lead to less precision.
- *
- * Please note that while canvas is more significantly more performant than DOM
- * measurement, there are subpixel to single digit pixel differences between
- * DOM and canvas measurement due to the different rendering engines used.
- */
-export class CanvasTextUtils {
-  context: CanvasRenderingContext2D;
-  currentText = '';
-
-  constructor({ font, container }: CanvasTextParams) {
-    this.context = document.createElement('canvas').getContext('2d')!;
-
-    // Set the canvas font to ensure text width calculations are correct
-    if (font) {
-      this.context.font = font;
-    } else if (container) {
-      this.context.font = this.computeFontFromElement(container);
-    }
-  }
-
-  computeFontFromElement = (element: HTMLElement) => {
-    const computedStyles = window.getComputedStyle(element);
-    // TODO: font-stretch is not included even though it potentially should be
-    // @see https://developer.mozilla.org/en-US/docs/Web/CSS/font#constituent_properties
-    // It appears to be unsupported and/or breaks font computation in canvas
-    return [
-      'font-style',
-      'font-variant',
-      'font-weight',
-      'font-size',
-      'font-family',
-    ]
-      .map((prop) => computedStyles.getPropertyValue(prop))
-      .join(' ')
-      .trim();
-  };
-
-  get textWidth() {
-    return this.context.measureText(this.currentText).width;
-  }
-
-  setTextToCheck = (text: string) => {
-    this.currentText = text;
-  };
-}
 
 /**
  * Utilities for truncating types at various positions, as well as
